Refactor code: create css file for highlights, rename images and add conctants
